move savedata to proper package
