+++ src/core/libraries/libs.cpp
@@ -17,7 +17,7 @@
 #include "core/libraries/system/commondialog.h"
 #include "core/libraries/system/msgdialog.h"
 #include "core/libraries/system/posix.h"
-#include "core/libraries/system/savedata.h"
+#include "core/libraries/save_data/save_data.h"
 #include "core/libraries/system/savedatadialog.h"
 #include "core/libraries/system/sysmodule.h"
 #include "core/libraries/system/systemservice.h"
